Extract event handling into separate class
